requirements.txt
strands-agents
strands-agents-tools 
strands-agents-builder
bedrock-agentcore